backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
google-genai
passlib
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
requests
